requirements.txt
django==4.2
djangorestframework==3.14
djangorestframework-camel-case==1.4.2
psycopg==3.1.9